+++ src/layouts/shared/force-directed-algorithm-iteration/index.ts
@@ -1,2 +1,2 @@
 export { ForceDirectedAlgorithmIteration } from "./force-directed-algorithm-iteration";
-export type { ForceDirectedAlgorithmIterationParams } from "./force-directed-algorithm-params";
+export type { ForceDirectedAlgorithmIterationParams } from "./force-directed-algorithm-iteration-params";
